feat(Layout): add component wrapper
