+++ src/Router.tsx
@@ -1,10 +1,17 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import { HelloDataverse } from './sections/hello-dataverse/HelloDataverse'
+import { Layout } from './sections/layout/Layout'
 
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <HelloDataverse />
+    element: <Layout />,
+    children: [
+      {
+        path: '/',
+        element: <HelloDataverse />
+      }
+    ]
   }
 ])
