+++ tests/sections/hello-dataverse/HelloDataverse.test.tsx
@@ -2,9 +2,11 @@ import { screen } from '@testing-library/react'
 import { HelloDataverse } from '../../../src/sections/hello-dataverse/HelloDataverse'
 import { renderWithRouter } from '../../renderWithRouter'
 
-test('renders hello dataverse title', () => {
-  renderWithRouter(<HelloDataverse />)
+describe('HelloDataverse page', () => {
+  it('renders hello dataverse title', () => {
+    renderWithRouter(<HelloDataverse />)
 
-  const titleElement = screen.getByRole('heading', { name: 'title' })
-  expect(titleElement).toBeInTheDocument()
+    const titleElement = screen.getByRole('heading', { name: 'title' })
+    expect(titleElement).toBeInTheDocument()
+  })
 })
